Added PUG Lookback
- Added array DS with serialize/deserialize operations
- Updated long-term to use array
- Some house keeping on where things are located:
	- monte-carlo
	- group-gen
	- etc.
- Added two new PUG Lookback specific flags in the ini file:
	- backtrack-direction
	- backtrack-traffic-count

+++ src/util/common.c
@@ -66,7 +66,7 @@ int dir_num_files(char const *dname) {
   return file_count;
 }
 
-int file_read(FILE *f, char **out) {
+size_t file_read(FILE *f, char **out) {
   size_t cap = 8192;
   size_t nbytes = 1024;
   size_t index = 0;
